requirements.txt
Django = 2.2
django-crispy-forms==1.12.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
httpie==2.5.0